# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music Catalog API",
        "description": "API для работы с каталогом исполнителей, альбомов и песен.",
        "contact": {
            "email": "admin@musiccatalog.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Получить список всех альбомов",
                "description": "Возвращает список всех альбомов в базе данных",
                "operationId": "735609843dcfb89ee4e12e340eddca25",
                "responses": {
                    "200": {
                        "description": "Успешный ответ с данными альбомов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Album"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Создать новый альбом",
                "description": "Создает новый альбом и возвращает его данные",
                "operationId": "2252aa8fd63b7d2fa1203c1735943d46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "artist_id",
                                    "release_year"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Название альбома",
                                        "type": "string"
                                    },
                                    "artist_id": {
                                        "description": "ID исполнителя",
                                        "type": "integer"
                                    },
                                    "release_year": {
                                        "description": "Год выпуска альбома",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Альбом успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Album"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
        "/api/albums/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Получить альбом по ID",
                "description": "Возвращает информацию об альбоме по его ID",
                "operationId": "71e9c4179582355e89419e6f6994c418",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID альбома",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о найденном альбоме",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Album"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Альбом не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Albums"
                ],
                "summary": "Обновить данные альбома",
                "description": "Обновляет данные альбома по его ID",
                "operationId": "1300ec85cc590b1ae1468c9bf8db122d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID альбома",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "artist_id",
                                    "release_year"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Название альбома",
                                        "type": "string"
                                    },
                                    "artist_id": {
                                        "description": "ID исполнителя",
                                        "type": "integer"
                                    },
                                    "release_year": {
                                        "description": "Год выпуска альбома",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Альбом успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Album"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Альбом не найден"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Удалить альбом",
                "description": "Удаляет альбом по его ID",
                "operationId": "cabfc7c152115607d310827d46a82d5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID альбома",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Альбом успешно удален"
                    },
                    "404": {
                        "description": "Альбом не найден"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Album": {
                "required": [
                    "title",
                    "artist_id",
                    "release_year"
                ],
                "properties": {
                    "id": {
                        "description": "ID альбома",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Название альбома",
                        "type": "string",
                        "example": "Abbey Road"
                    },
                    "artist_id": {
                        "description": "ID исполнителя",
                        "type": "integer",
                        "example": 1
                    },
                    "release_year": {
                        "description": "Год выпуска альбома",
                        "type": "integer",
                        "example": 1969
                    },
                    "created_at": {
                        "description": "Дата создания записи",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-01T14:30:00Z"
                    },
                    "updated_at": {
                        "description": "Дата последнего обновления записи",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-01T14:30:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}